[Task-2-1] Moved comp_op to pkg
