+++ vadim.nesmeianov/task-2-1/pkg/comp_op/comp_op.go
